elevator-related subsystems are created in elevator class
