+++ src/main/java/frc/robot/auto/RedNoProcess3PieceAuto.java
@@ -8,18 +8,18 @@ import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.drivetrain.Drivetrain;
 import frc.robot.elevator.CoralRoller;
-import frc.robot.elevator.ElevatorCGs;
+import frc.robot.elevator.Elevator;
 import java.util.Optional;
 
 public class RedNoProcess3PieceAuto extends AutoMode {
 
   CoralRoller coralRoller;
-  ElevatorCGs elevatorCG;
+  Elevator elevator;
 
-  public RedNoProcess3PieceAuto(Drivetrain drivetrain, ElevatorCGs autoCommandGroups) {
+  public RedNoProcess3PieceAuto(Drivetrain drivetrain, Elevator elevatorSystem) {
     super(drivetrain);
-    elevatorCG = autoCommandGroups;
-    coralRoller = elevatorCG.getCoralRoller();
+    elevator = elevatorSystem;
+    coralRoller = elevator.getCoralRoller();
   }
 
   AutoRoutine redNoProcess3PieceRoutine =
@@ -46,7 +46,7 @@ public class RedNoProcess3PieceAuto extends AutoMode {
 
     redNoProcess3PieceRoutine
         .active()
-        .onTrue(Commands.parallel(redNoProcessToL4I.cmd(), elevatorCG.coralL4PositionCommand()));
+        .onTrue(Commands.parallel(redNoProcessToL4I.cmd(), elevator.coralL4PositionCommand()));
 
     redNoProcessToL4I
         .done()
@@ -56,14 +56,14 @@ public class RedNoProcess3PieceAuto extends AutoMode {
                 coralRoller.createOuttakeCommand(),
                 new WaitCommand(0.2),
                 coralRoller.createStopCommand(),
-                new ParallelCommandGroup(redL4IToSource.cmd(), elevatorCG.coralIntakeCommand())));
+                new ParallelCommandGroup(redL4IToSource.cmd(), elevator.coralIntakeCommand())));
 
     redL4IToSource
         .done()
         .onTrue(
             Commands.sequence(
                 new WaitCommand(0.2),
-                Commands.parallel(redSourceToL4L.cmd(), elevatorCG.coralL4PositionCommand())));
+                Commands.parallel(redSourceToL4L.cmd(), elevator.coralL4PositionCommand())));
 
     redSourceToL4L
         .done()
@@ -73,14 +73,14 @@ public class RedNoProcess3PieceAuto extends AutoMode {
                 coralRoller.createOuttakeCommand(),
                 new WaitCommand(0.2),
                 coralRoller.createStopCommand(),
-                Commands.parallel(elevatorCG.coralIntakeCommand(), redL4LToSource.cmd())));
+                Commands.parallel(elevator.coralIntakeCommand(), redL4LToSource.cmd())));
 
     redL4LToSource
         .done()
         .onTrue(
             Commands.sequence(
                 new WaitCommand(0.2),
-                Commands.parallel(elevatorCG.coralL4PositionCommand(), redSourceToL4K.cmd())));
+                Commands.parallel(elevator.coralL4PositionCommand(), redSourceToL4K.cmd())));
 
     redSourceToL4K
         .done()
